requirements.txt
matplotlib
librosa
tqdm
pandas
numba
numpy
scipy
PyYAML
tensorboardX
pyloudnorm
setuptools
g2p_en
resemblyzer
webrtcvad
tensorboard
scikit-learn
scikit-image
textgrid
jiwer
pycwt
PyWavelets
praat-parselmouth==0.3.3
jieba
einops
chardet
pretty-midi
pytorch-lightning==0.7.1
torchvision
h5py
pypinyin
g2pM
git+https://github.com/sooftware/conformer
git+https://github.com/CompVis/taming-transformers
